Give the calendar's day arithmetic a module of its own
