+++ modules/event-list/index.ts
@@ -3,5 +3,6 @@ export type {CalendarBodyHandle, CalendarSource, PoweredBy, SourcedEvent} from '
 export * as EventList from './event-list'
 export * as EventDetail from './event-detail'
 export {shareEvent, eventKey} from './calendar-util'
+export {daysWithEvents, deriveDays, eventsOnDay, occursOn} from './days'
 export {timelineBlocks, timelineEntries, timelineWindow} from './timeline'
 export type {TimelineBlock, TimelineWindow} from './timeline'
